Km451: Move github(rest) to github_direct
